requirements.txt
playwright>=1.44.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.0
lxml>=5.2.0
httpx>=0.27.0
pyyaml>=6.0.1
python-dateutil>=2.9.0
